clip-service/requirements.txt
fastapi
uvicorn[standard]
Pillow
requests
torch
numpy
open-clip-torch
